refactor(installer): add view model for the top-level wizard
